+++ FuzzySharp/SimilarityRatio/Scorer/Generic/IRatioScorer.cs
@@ -4,6 +4,6 @@ namespace FuzzySharp.SimilarityRatio.Scorer.Generic
 {
     public interface IRatioScorer<in T> where T : IEquatable<T>
     {
-        int Score(T[] input1, T[] input2);
+        double Score(T[] input1, T[] input2);
     }
 }
