+++ FuzzySharp/SimilarityRatio/Scorer/Generic/ScorerBase.cs
@@ -4,6 +4,6 @@ namespace FuzzySharp.SimilarityRatio.Scorer.Generic
 {
     public abstract class ScorerBase<T> : IRatioScorer<T> where T : IEquatable<T>
     {
-        public abstract int Score(T[] input1, T[] input2);
+        public abstract double Score(T[] input1, T[] input2);
     }
 }
