+++ FuzzySharp/SimilarityRatio/Scorer/IRatioScorer.cs
@@ -4,7 +4,7 @@ namespace FuzzySharp.SimilarityRatio.Scorer
 {
     public interface IRatioScorer
     {
-        int Score(string input1, string input2);
-        int Score(string input1, string input2, PreprocessMode preprocessMode);
+        double Score(string input1, string input2);
+        double Score(string input1, string input2, PreprocessMode preprocessMode);
     }
 }
