+++ FuzzySharp/SimilarityRatio/Scorer/ScorerBase.cs
@@ -4,9 +4,9 @@ namespace FuzzySharp.SimilarityRatio.Scorer
 {
     public abstract class ScorerBase : IRatioScorer
     {
-        public abstract int Score(string input1, string input2);
+        public abstract double Score(string input1, string input2);
 
-        public int Score(string input1, string input2, PreprocessMode preprocessMode)
+        public double Score(string input1, string input2, PreprocessMode preprocessMode)
         {
             var preprocessor = StringPreprocessorFactory.GetPreprocessor(preprocessMode);
             input1 = preprocessor(input1);
